% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_Graphic.R
\name{cat_smd}
\alias{cat_smd}
\title{smd value for categorical variables}
\usage{
cat_smd(ntable, var, data)
}
\arguments{
\item{ntable}{propotion table of baseline categorical variable and group variable}

\item{var}{baseline categorical variable}

\item{data}{data}
}
\description{
smd value for categorical variables
}
\examples{
set.seed(2018)
group <-round(abs(rnorm(500)*10),0) \%\% 2
cont_1 <-round(abs(rnorm(500)*10),0)
cat_multi_1 <-round(abs(rnorm(500)*10),0) \%\% 3
data_check <-data.frame(group, cont_1, cat_multi_1)
data_check$group <- factor(data_check$group, levels = c(0,1), labels = c("Control","Treatment"))
data_check$cat_multi_1 <- factor(data_check$cat_multi_1)
cat_smd(table(data_check$cat_multi_1, data_check$group),"cat_multi_1",data_check )
}
