% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_Graphic.R
\name{cont_smd}
\alias{cont_smd}
\title{smd value for continuous variable.}
\usage{
cont_smd(mean1, mean2, var1, var2)
}
\arguments{
\item{mean1}{mean of a baseline variable in the treatment group.}

\item{mean2}{mean of a baseline variable in the control group.}

\item{var1}{variance a baseline variable in the treatment group.}

\item{var2}{variance of a baseline variable in the control group.}
}
\value{
smd value
}
\description{
smd value for continuous variable.
}
\examples{
cont_smd(10,11,2,3)
}
