% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_Graphic.R
\name{get_mean}
\alias{get_mean}
\title{Mean, var function}
\usage{
get_mean(x)
}
\arguments{
\item{x}{variable}
}
\value{
mean table
}
\description{
Mean, var function
}
\examples{
get_mean(round(abs(rnorm(500)*10),0))
}
