% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_Graphic.R
\name{mydocx}
\alias{mydocx}
\title{write smd table or demographic table into docx file}
\usage{
mydocx(smd_table, name)
}
\arguments{
\item{smd_table}{smd table or demo graphic table.}

\item{name}{file name to save}
}
\description{
write smd table or demographic table into docx file
}
\examples{
mydocx(data.frame(smd.value <- 3.4, smd.lo <- 1.1, smd.up <- 5.6),"smd_table")
}
