% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_Graphic.R
\name{smd_ci}
\alias{smd_ci}
\title{Confident interval for smd}
\usage{
smd_ci(n1, n2, smd)
}
\arguments{
\item{n1}{length of a baseline variable in the treatment group.}

\item{n2}{length of a baseline variable in the control group.}

\item{smd}{smd value}
}
\value{
vector of 95% ci
}
\description{
Confident interval for smd
}
\examples{
smd_ci(10,12,0.3)
}
