% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_Graphic.R
\name{t.test.p.value}
\alias{t.test.p.value}
\title{t.test to calculate p value}
\usage{
\method{t}{test.p.value}(...)
}
\arguments{
\item{...}{variables}
}
\value{
p value
}
\description{
t.test to calculate p value
}
